Cleanup code per Carlos's comments
